refactor: Move utils package functions to staking app. update next-config + imports.
