+++ apps/staking/next.config.js
@@ -1,12 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-disable no-undef */
 /** @type {import('next').NextConfig} */
-const modules = [
-    '@explorer/ui',
-    '@explorer/wallet',
-    '@explorer/utils',
-    '@explorer/services',
-];
+const modules = ['@explorer/ui', '@explorer/wallet', '@explorer/services'];
 
 const ContentSecurityPolicy = `
   default-src 'self';
