--- packages/utils/.eslintrc
@@ -1,4 +1,0 @@
-{
-    "root": true,
-    "extends": ["custom"]
-}
