--- packages/utils/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "@explorer/utils",
-    "version": "0.0.0",
-    "main": "./src/index.tsx",
-    "types": "./src/index.tsx",
-    "scripts": {
-        "lint": "TIMING=1 eslint --ext .tsx,.ts src/"
-    },
-    "peerDependencies": {
-        "react": "18.2.0",
-        "react-dom": "18.2.0",
-        "lodash": "4.17.21"
-    },
-    "devDependencies": {
-        "eslint": "8.44.0",
-        "eslint-config-custom": "*",
-        "tsconfig": "*",
-        "typescript": "5.7.3",
-        "@explorer/prettier-config": "*",
-        "@types/react": "18.3.12",
-        "@types/react-dom": "18.3.1"
-    },
-    "dependencies": {}
-}
