--- packages/utils/prettier.config.js
@@ -1,4 +1,0 @@
-/* eslint-disable no-undef */
-module.exports = {
-    ...require('@explorer/prettier-config'),
-};
