--- packages/utils/tsconfig.json
@@ -1,5 +1,0 @@
-{
-    "extends": "tsconfig/react-library.json",
-    "include": ["."],
-    "exclude": ["dist", "build", "node_modules"]
-}
